Update lobby.sql
privacy check eligibity of requests
can do that to owner ?

+++ _bin/sql/func/lobby.sql
@@ -57,6 +57,8 @@ UTILS
 	lobby_utils_refresh_member_authz
 	lobby_utils_delete_member_invitation
 */
+--todo refresh eligibility waiting_request / in set_privacy, set_owner
+--todo change order of execution lock request and invitations using for key skip locked
 
 DROP FUNCTION IF EXISTS lobby_create, lobby_join, lobby_leave,
   lobby_user_joinrequest_deny, lobby_manage_joinrequest_accept, lobby_manage_joinrequest_deny,
@@ -80,34 +82,34 @@ END;
 $$ LANGUAGE plpgsql;
 
 CREATE OR REPLACE FUNCTION lobby_join(_id_viewer integer, _id_lobby integer) RETURNS integer AS $$
+/*
+RETURN
+	1 - user joined lobby
+	2 - user created request
+*/
 DECLARE
-  __lobby_params lobbys%rowtype;
+  __lobby_params record;
   
   __viewer_perms integer;
   __allowed_to_join boolean := FALSE;
   __has_joined boolean := FALSE;
 BEGIN
-  SELECT * INTO __lobby_params FROM lobbys WHERE id=_id_lobby FOR SHARE;
+  SELECT check_join, privacy, id_owner, authz_default, authz_follow, authz_friend INTO __lobby_params FROM lobbys WHERE id=_id_lobby FOR SHARE;
   IF NOT FOUND THEN RAISE EXCEPTION 'lobby not found'; END IF;
 
-  --user
-  PERFORM pg_advisory_xact_lock_shared(hashtextextended('user_user:'||least(_id_viewer, __lobby_params.id_owner),greatest(_id_viewer, __lobby_params.id_owner)));
-  PERFORM FROM user_bans WHERE id_usera=least(_id_viewer, __lobby_params.id_owner) AND id_userb=greatest(_id_viewer, __lobby_params.id_owner) FOR KEY SHARE;
-  IF FOUND THEN RAISE EXCEPTION 'users_ban'; END IF;
-	PERFORM FROM friends WHERE id_usera=least(_id_viewer, __lobby_params.id_owner) AND id_userb=greatest(_id_viewer, __lobby_params.id_owner) FOR KEY SHARE;
-  IF FOUND THEN
-    __viewer_perms := __lobby_params.authz_friend;
-    __allowed_to_join := __lobby_params.privacy IN('FRIEND', 'FOLLOWER', 'DEFAULT');
-  ELSE
-    PERFORM FROM follows WHERE id_follower=_id_viewer AND id_following=__lobby_params.id_owner FOR KEY SHARE;
-    IF FOUND THEN
+  --user_user rel
+  CASE relation_user_user(_id_viewer, __lobby_params.id_owner)
+    WHEN -1 THEN RAISE EXCEPTION 'user_user_ban';
+    WHEN 3 THEN
+      __viewer_perms := __lobby_params.authz_friend;
+      __allowed_to_join := __lobby_params.privacy IN('FRIEND', 'FOLLOWER', 'DEFAULT');
+    WHEN 2 THEN
       __viewer_perms := __lobby_params.authz_follow;
       __allowed_to_join := __lobby_params.privacy IN('FOLLOWER', 'DEFAULT');
     ELSE
       __viewer_perms := __lobby_params.authz_default;
       __allowed_to_join := __lobby_params.privacy IN('DEFAULT');
-    END IF;
-  END IF;
+  END CASE;
 
   IF __allowed_to_join THEN
 		INSERT INTO lobby_users (id_user, id_lobby, fk_member, is_owner, authz, joinrequest_status)
@@ -197,8 +199,7 @@ BEGIN
   UPDATE lobby_users SET joinrequest_status=CASE joinrequest_status
       WHEN 'WAITING_LOBBY' THEN 'WAITING_USER'::lobby_active_joinrequest_status
       ELSE 'INV_WAITING_USER'::lobby_active_joinrequest_status END
-    WHERE id_lobby=_id_lobby
-      AND id_user=_id_target
+    WHERE id_lobby=_id_lobby AND id_user=_id_target
       AND joinrequest_status IN('WAITING_LOBBY', 'INV_WAITING_LOBBY');
 	IF NOT FOUND THEN RAISE EXCEPTION 'lobby_joinrequest not found'; END IF;
 
@@ -228,12 +229,12 @@ DECLARE
 BEGIN
   PERFORM FROM friends WHERE id_usera=least(_id_viewer, _id_target) AND id_userb=greatest(_id_viewer, _id_target) FOR KEY SHARE;
   IF NOT FOUND THEN RAISE EXCEPTION 'not friends'; END IF;
+
   SELECT check_join INTO __check_join FROM lobbys WHERE id=_id_lobby FOR SHARE;
   IF NOT FOUND THEN RAISE EXCEPTION 'lobby not found'; END IF;
   SELECT (0 < authz OR __check_join IS FALSE) INTO __trust_invite FROM lobby_users WHERE fk_member=_id_viewer AND id_lobby=_id_lobby FOR SHARE;
   IF NOT FOUND THEN RAISE EXCEPTION 'user not lobby_member'; END IF;
 
-  --always update if allowed to add lobby_invitations
   INSERT INTO lobby_users(id_user, id_lobby, joinrequest_status)
     VALUES(_id_target,
            _id_lobby,
@@ -273,7 +274,7 @@ $$ LANGUAGE plpgsql;
 
 CREATE OR REPLACE FUNCTION lobby_set_check_join(_id_viewer integer, _id_lobby integer, _check_join boolean) RETURNS boolean AS $$
 BEGIN
-  UPDATE lobbys SET check_join=_check_join --prevent add lobby_invitations/lobby_users
+  UPDATE lobbys SET check_join=_check_join
     WHERE id=_id_lobby
       AND id_owner=_id_viewer
       AND check_join<>_check_join;
@@ -300,7 +301,37 @@ CREATE OR REPLACE FUNCTION lobby_set_privacy(_id_viewer integer, _id_lobby integ
 BEGIN
   UPDATE lobbys SET privacy=_privacy WHERE id_owner=_id_viewer AND id=_id_lobby AND privacy<>_privacy;
   IF NOT FOUND THEN RAISE EXCEPTION 'update failed'; END IF;
-	--refresh request eligibility if check_join is true
+
+  --todo use this when owner change ?
+  IF _privacy<>'DEFAULT' THEN --test
+    --update requests to invitation, delete requests
+	  PERFORM FROM lobby_users WHERE id_lobby=_id_lobby AND joinrequest_status IS NOT NULL ORDER BY id_user FOR UPDATE;
+
+	  IF _privacy='PRIVATE' THEN --update request to invitations and delete requests
+      UPDATE lobby_users SET joinrequest_status=CASE WHEN joinrequest_status='WAITING_LOBBY' THEN 'INV_WAITING_LOBBY' ELSE 'INV_WAITING_USER' END
+       WHERE id_lobby=_id_lobby AND joinrequest_status IN('WAITING_LOBBY', 'WAITING_USER')
+         AND EXISTS(SELECT FROM lobby_invitations WHERE id_target=lobby_users.id_user AND id_lobby=_id_lobby FOR KEY SHARE);
+      DELETE FROM lobby_users WHERE joinrequest_status IN('WAITING_LOBBY','WAITING_USER') AND id_lobby=_id_lobby;
+    ELSE
+      --delete request not eligible without invitations
+      DELETE FROM lobby_users
+        WHERE joinrequest_status IN('WAITING_LOBBY','WAITING_USER') AND id_lobby=_id_lobby
+	        AND (CASE relation_user_user(lobby_users.id_user, _id_viewer)
+	            WHEN 3 THEN _privacy<>'PRIVATE'
+	            WHEN 2 THEN _privacy IN('DEFAULT', 'FOLLOW')
+	            ELSE _privacy IN('DEFAULT') END) IS NOT TRUE
+	        AND NOT EXISTS(SELECT FROM lobby_invitations WHERE id_target=lobby_users.id_user AND id_lobby=_id_lobby FOR KEY SHARE);
+
+			--update request not eligible with invitation to invitation
+      UPDATE lobby_users SET joinrequest_status=CASE WHEN joinrequest_status='WAITING_LOBBY' THEN 'INV_WAITING_LOBBY' ELSE 'INV_WAITING_USER' END
+        WHERE id_lobby=_id_lobby AND joinrequest_status IN('WAITING_LOBBY', 'WAITING_USER')
+          AND (CASE relation_user_user(lobby_users.id_user, _id_viewer)
+	            WHEN 3 THEN _privacy<>'PRIVATE'
+	            WHEN 2 THEN _privacy IN('DEFAULT', 'FOLLOW')
+	            ELSE _privacy IN('DEFAULT') END) IS NOT TRUE
+          AND EXISTS(SELECT FROM lobby_invitations WHERE id_target=lobby_users.id_user AND id_lobby=_id_lobby FOR KEY SHARE);
+    END IF;
+  END IF;
   RETURN true;
 END
 $$ LANGUAGE plpgsql;
@@ -341,7 +372,6 @@ BEGIN
 	IF NOT FOUND THEN RAISE EXCEPTION 'lobby_users target not found'; END IF;
 
   PERFORM FROM lobby_utils_refresh_member_authz(_id_lobby, true);
-  
   RETURN true;
 END
 $$ LANGUAGE plpgsql;
@@ -350,10 +380,6 @@ CREATE OR REPLACE FUNCTION lobby_ban_user(_id_viewer integer, _id_lobby integer,
 DECLARE
   __was_member boolean;
 BEGIN
-  IF(_id_viewer=_id_target)
-    THEN RAISE EXCEPTION '_id_viewer == _id_target';
-  END IF;
-
   PERFORM FROM lobby_users WHERE id_user=_id_viewer AND id_lobby=_id_lobby AND is_owner FOR SHARE;
 	IF NOT FOUND THEN RAISE EXCEPTION 'lobby_user unauthz'; END IF;
 
@@ -387,12 +413,12 @@ DECLARE
   __change_slots integer;
   __ids_member integer[];
 BEGIN
-	PERFORM FROM lobbys WHERE id=_id_lobby AND id_owner=_id_viewer FOR UPDATE; --prevent join
+	PERFORM FROM lobbys WHERE id=_id_lobby AND id_owner=_id_viewer FOR NO KEY UPDATE; --prevent join
 	IF NOT FOUND THEN RAISE EXCEPTION 'lobby not found'; END IF;
 
 	SELECT _max_slots-(max_slots-free_slots) INTO __change_slots
 		FROM lobby_slots
-		WHERE id_lobby=_id_lobby AND _max_slots<>max_slots FOR UPDATE; --prevent left
+		WHERE id_lobby=_id_lobby AND _max_slots<>max_slots FOR NO KEY UPDATE; --prevent left
 	IF NOT FOUND THEN RAISE EXCEPTION 'lobby_slots not found'; END IF;
 
 	UPDATE lobby_slots SET max_slots=_max_slots,
@@ -422,55 +448,55 @@ DROP FUNCTION IF EXISTS lobby_utils_refresh_member_authz, lobby_utils_delete_mem
 
 CREATE OR REPLACE FUNCTION lobby_utils_refresh_member_authz(_id_lobby integer,  _refresh_owner boolean) RETURNS void AS $$
 /*
-update members.authz
 lobby.check_join ? update members.invitations
 */
 DECLARE
   __authz integer[];
   __id_owner integer;
   __check_join boolean;
-  __ids_request integer[];
+  __ids_request_creator integer[];
 BEGIN
   SELECT ARRAY[authz_default, authz_follow, authz_friend], id_owner, check_join INTO __authz, __id_owner,__check_join FROM lobbys WHERE lobbys.id=_id_lobby FOR UPDATE;
   IF __check_join IS FALSE THEN
 	  UPDATE lobby_users
 			SET authz =
-				CASE WHEN EXISTS(SELECT pg_advisory_xact_lock_shared(hashtextextended('user_user:' || least(fk_member, __id_owner),greatest(fk_member, __id_owner)))) THEN
-					CASE WHEN EXISTS(SELECT FROM friends WHERE id_usera = least(__id_owner, fk_member) AND id_userb = greatest(__id_owner, fk_member) FOR KEY SHARE) THEN __authz[1]
-						WHEN EXISTS(SELECT FROM follows WHERE id_follower = fk_member AND id_following = __id_owner FOR KEY SHARE) THEN __authz[2]
-						ELSE __authz[3]
-					END
+				CASE relation_user_user(lobby_users.id_user, __id_owner) --allowed_request
+	        WHEN 3 THEN authz[3]
+	        WHEN 2 THEN authz[2]
+	        ELSE authz[1]
 				END
 		  WHERE fk_member IS NOT NULL AND id_lobby=_id_lobby AND is_owner IS FALSE;
 	ELSE
-    WITH upd_members AS (
+    WITH upd_members AS ( --lock lobby_creators
 			UPDATE lobby_users
-		    SET authz = CASE WHEN is_owner THEN 1
-          WHEN EXISTS(SELECT pg_advisory_xact_lock_shared(hashtextextended('user_user:' || least(fk_member, __id_owner),greatest(fk_member, __id_owner)))) THEN
-					  CASE WHEN EXISTS(SELECT FROM friends WHERE id_usera = least(__id_owner, fk_member) AND id_userb = greatest(__id_owner, fk_member) FOR KEY SHARE) THEN __authz[1]
-						  WHEN EXISTS(SELECT FROM follows WHERE id_follower = fk_member AND id_following = __id_owner FOR KEY SHARE) THEN __authz[2]
-						  ELSE __authz[3]
-					  END
-				  END
+		    SET authz =
+		      CASE WHEN is_owner THEN 1
+		      ELSE
+            CASE relation_user_user(lobby_users.id_user, __id_owner) --allowed_request
+	            WHEN 3 THEN authz[3]
+	            WHEN 2 THEN authz[2]
+	          ELSE authz[1] END
+		      END
         WHERE fk_member IS NOT NULL AND id_lobby=_id_lobby AND ((is_owner AND _refresh_owner) OR is_owner IS FALSE)
         RETURNING id_user, authz
     )
-    SELECT array_agg(t_target_users.id) INTO __ids_request FROM (
-      SELECT lobby_users.id_user AS id
+    SELECT array_agg(t_target_creators.id) INTO __ids_request_creator FROM ( --lock lobby_targets
+      SELECT t_invitations.id_creator AS id
         FROM (
-			    SELECT DISTINCT lobby_invitations.id_target AS id_target FROM (SELECT id_user FROM upd_members WHERE upd_members.authz > 0) t_creators
-	          JOIN lobby_invitations ON lobby_invitations.id_creator=t_creators.id_user AND lobby_invitations.id_lobby=_id_lobby
-	        ) t_targets
-	      JOIN lobby_users ON lobby_users.id_user=t_targets.id_target
-				WHERE lobby_users.joinrequest_status IS NOT NULL --NOT IN('INV_WAITING_USER', 'WAITING_USER') --issue #2
-        ORDER BY id_user FOR UPDATE OF lobby_users) t_target_users;
-    
-    UPDATE lobby_users --todo get id_actor
+			    SELECT li.id_target, li.id_creator FROM (SELECT id_user FROM upd_members WHERE upd_members.authz > 0) t_upd_members
+	          JOIN lobby_invitations li ON li.id_creator=t_upd_members.id_user AND li.id_lobby=_id_lobby
+	        ) t_invitations
+	      JOIN lobby_users ON lobby_users.id_user=t_invitations.id_target
+				WHERE lobby_users=_id_lobby AND lobby_users.joinrequest_status IS NOT NULL --NOT IN('INV_WAITING_USER', 'WAITING_USER') --issue #2
+        ORDER BY id_user FOR UPDATE OF lobby_users) t_target_creators;
+    --test
+    UPDATE lobby_users lu
       SET joinrequest_status=
         CASE WHEN joinrequest_status='WAITING_LOBBY'
           THEN 'WAITING_USER'::lobby_active_joinrequest_status
           ELSE 'INV_WAITING_USER'::lobby_active_joinrequest_status END
-      WHERE id_user=ANY(__ids_request) AND id_lobby=_id_lobby AND joinrequest_status NOT IN('INV_WAITING_USER', 'WAITING_USER');
+      FROM (SELECT id_target, id_creator FROM lobby_invitations WHERE id_creator=ANY(__ids_request_creator) AND id_lobby=_id_lobby FOR KEY SHARE) li
+      WHERE li.id_target=lu.id_user AND lu.id_lobby=_id_lobby AND lu.joinrequest_status NOT IN('INV_WAITING_USER', 'WAITING_USER');
 	END IF;
 END;
 $$ LANGUAGE plpgsql;
@@ -489,14 +515,43 @@ BEGIN
       ORDER BY id_user FOR UPDATE OF lobby_users;
 
   WITH del_lobby_invitations AS (
-    DELETE FROM lobby_invitations WHERE id_creator=ANY(_ids_member) AND id_lobby=_id_lobby RETURNING id_target
+    DELETE FROM lobby_invitations WHERE id_creator=ANY(_ids_member) AND id_lobby=_id_lobby RETURNING id_target, id_creator
 	)
 	SELECT array_agg(id_target) INTO __ids_target FROM del_lobby_invitations;
   
-  DELETE FROM lobby_users
+  DELETE FROM lobby_users --todo get id_actor
     WHERE id_user=ANY(__ids_target)
       AND id_lobby=_id_lobby
       AND joinrequest_status NOT IN('WAITING_USER', 'WAITING_LOBBY')
       AND NOT EXISTS(SELECT FROM lobby_invitations WHERE id_target=lobby_users.id_user AND id_lobby=_id_lobby FOR KEY SHARE);--SKIP LOCKED
 END;
+$$ LANGUAGE plpgsql;
+
+CREATE OR REPLACE FUNCTION relation_user_user(_id_viewer integer, _id_target integer) RETURNS integer AS $$
+/*
+RETURN
+  0 = _id_viewer=_id_target
+	1 = no relation
+	2 = _id_viewer follow _id_target
+	3 = are friends
+	-1= block
+*/
+DECLARE
+  __least integer DEFAULT least(_id_viewer, _id_target);
+  __greatest integer DEFAULT greatest(_id_viewer, _id_target);
+BEGIN
+  IF(_id_viewer=_id_target) THEN RAISE EXCEPTION '_id_viewer=_id_target'; END IF;
+	--test
+	PERFORM pg_advisory_xact_lock_shared(hashtextextended('user_user:' || __least,__greatest));
+	PERFORM FROM user_bans WHERE id_usera=__least AND id_userb=__greatest FOR KEY SHARE;
+	IF FOUND THEN RETURN -1; END IF;
+
+	PERFORM FROM friends WHERE id_usera=__least AND id_userb=__greatest FOR KEY SHARE;
+	IF FOUND THEN RETURN 3; END IF;
+
+	PERFORM FROM follows WHERE id_follower=_id_viewer AND id_following=_id_target FOR KEY SHARE;
+	IF FOUND THEN RETURN 2;
+	ELSE RETURN 1;
+	END IF;
+END
 $$ LANGUAGE plpgsql;
